partially implemented random encounters
I added a function that implements a basic random encounter. Currently
it activates after 5 seconds and then deactivates after 10. It doesn't
allow any movement of the character and it doesn't allow the timer I'm
using to update while menu is up and doesn't allow the user to open menu
when it is active.